add new

+++ src/main/java/com/lizan/community/controller/publishController.java
@@ -1,6 +1,5 @@
 package com.lizan.community.controller;
 
-import com.lizan.community.Mapper.QuestionMapper;
 import com.lizan.community.dto.QuestionDto;
 import com.lizan.community.model.Question;
 import com.lizan.community.model.User;
@@ -23,7 +22,7 @@ public class publishController {
 
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id")Long id,
                        Model model) {
         QuestionDto question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -43,7 +42,7 @@ public class publishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
